Added frmEditStock module

+++ ShopManageSystem/ShopManageSystem/frmStockManagement.vb
@@ -33,34 +33,39 @@ Public Class frmStockManagement
     End Sub
 
     Private Sub btnDelete_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btnDelete.Click
-        Dim Rows As New ArrayList()
-        ' Add the selected rows to the list.
-        Dim I As Integer
-        For I = 0 To StockGV.SelectedRowsCount() - 1
-            If (StockGV.GetSelectedRows()(I) >= 0) Then
-                Rows.Add(StockGV.GetDataRow(StockGV.GetSelectedRows()(I)))
-            End If
-        Next
+        If StockGV.SelectedRowsCount > 0 Then
+            Dim Rows As New ArrayList()
+            ' Add the selected rows to the list.
+            Dim I As Integer
+            For I = 0 To StockGV.SelectedRowsCount() - 1
+                If (StockGV.GetSelectedRows()(I) >= 0) Then
+                    Rows.Add(StockGV.GetDataRow(StockGV.GetSelectedRows()(I)))
+                End If
+            Next
 
-        For I = 0 To Rows.Count - 1
-            Dim Row As DataRow = CType(Rows(I), DataRow)
+            For I = 0 To Rows.Count - 1
+                Dim Row As DataRow = CType(Rows(I), DataRow)
 
-            Dim deleteStockProduct As New OleDbCommand("DELETE FROM tblProduct WHERE prod_id = ?", conn)
-            deleteStockProduct.Parameters.AddWithValue("prod_id", Row("prod_id"))
+                Dim deleteStockProduct As New OleDbCommand("DELETE FROM tblProduct WHERE prod_id = ?", conn)
+                deleteStockProduct.Parameters.AddWithValue("prod_id", Row("prod_id"))
 
-            Try
-                conn.Open()
-                deleteStockProduct.ExecuteNonQuery()
-            Catch ex As Exception
-                MsgBox(ex.Message)
-            Finally
-                conn.Close()
-                deleteStockProduct.Parameters.Clear()
-            End Try
-        Next
+                Try
+                    conn.Open()
+                    deleteStockProduct.ExecuteNonQuery()
+                Catch ex As Exception
+                    MsgBox(ex.Message)
+                Finally
+                    conn.Close()
+                    deleteStockProduct.Parameters.Clear()
+                End Try
+            Next
+            XtraMessageBox.Show("The selected data has been deleted", "Data Deleted", MessageBoxButtons.OK, MessageBoxIcon.Asterisk)
 
-        'finally remove the rows from grid control, so that no repopulation is needed
-        StockGV.DeleteSelectedRows()
+            'finally remove the rows from grid control, so that no repopulation is needed
+            StockGV.DeleteSelectedRows()
+        Else
+            XtraMessageBox.Show("Please select a record to perform this action!", "No data selected", MessageBoxButtons.OK, MessageBoxIcon.Exclamation)
+        End If
     End Sub
 
     Public Sub fillStockGV()
@@ -70,10 +75,10 @@ Public Class frmStockManagement
             Dim da As OleDbDataAdapter 'define it here so it escape the undefine error below
 
             If HideZeroStockState = 0 Then
-                da = New OleDbDataAdapter("SELECT tblProduct.prod_id, tblProduct.prod_model, tblCategory.cat_name, tblProduct.prod_quantity, tblProduct.prod_cost, tblProduct.prod_price, tblProduct.prod_description" _
+                da = New OleDbDataAdapter("SELECT tblProduct.prod_id, tblProduct.prod_model, tblCategory.cat_name, tblProduct.prod_quantity, tblProduct.prod_cost, tblProduct.prod_price, tblProduct.prod_description, tblProduct.prod_category" _
                                                 & " FROM tblCategory INNER JOIN tblProduct ON tblCategory.cat_id = tblProduct.prod_category", openConn())
             Else
-                da = New OleDbDataAdapter("SELECT tblProduct.prod_id, tblProduct.prod_model, tblCategory.cat_name, tblProduct.prod_quantity, tblProduct.prod_cost, tblProduct.prod_price, tblProduct.prod_description" _
+                da = New OleDbDataAdapter("SELECT tblProduct.prod_id, tblProduct.prod_model, tblCategory.cat_name, tblProduct.prod_quantity, tblProduct.prod_cost, tblProduct.prod_price, tblProduct.prod_description, tblProduct.prod_category" _
                                                 & " FROM tblCategory INNER JOIN tblProduct ON tblCategory.cat_id = tblProduct.prod_category WHERE tblProduct.prod_quantity > 0", openConn())
             End If
 
@@ -112,6 +117,8 @@ Public Class frmStockManagement
             StockGV.Columns(5).DisplayFormat.FormatString = "c2"
 
             StockGV.Columns(6).Visible = False 'prod_description
+
+            StockGV.Columns(7).Visible = False 'prod_category
         Catch ex As Exception
             MsgBox(ex.Message)
         End Try
@@ -159,7 +166,7 @@ Public Class frmStockManagement
             'finally refresh gridview
             fillStockGV()
         Else
-            XtraMessageBox.Show("Please select a record to perform this action!", "No data selected", MessageBoxButtons.OK, MessageBoxIcon.Asterisk)
+            XtraMessageBox.Show("Please select a record to perform this action!", "No data selected", MessageBoxButtons.OK, MessageBoxIcon.Exclamation)
         End If
     End Sub
 
@@ -173,6 +180,24 @@ Public Class frmStockManagement
 
     Private Sub btnNew_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btnNew.Click
         frmAddStock.ShowDialog()
+    End Sub
+
+    Private Sub btnEdit_Click(ByVal sender As System.Object, ByVal e As System.EventArgs) Handles btnEdit.Click
+        If StockGV.SelectedRowsCount > 1 Then
+            XtraMessageBox.Show("You may only select one row of data!", "More than one data selected", MessageBoxButtons.OK, MessageBoxIcon.Exclamation)
+        ElseIf StockGV.SelectedRowsCount = 0 Then
+            XtraMessageBox.Show("Please select a record to perform this action!", "No data selected", MessageBoxButtons.OK, MessageBoxIcon.Exclamation)
+        Else
+            frmEditStock.CurrentProdID = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_id")
+            frmEditStock.txtModel.Text = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_model")
+            frmEditStock.txtQuantity.Text = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_quantity")
+            frmEditStock.txtCost.Text = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_cost")
+            frmEditStock.txtPrice.Text = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_price")
+            frmEditStock.txtDescription.Text = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_description")
+            frmEditStock.pceDDCategory.Text = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "cat_name")
+            frmEditStock.CurrentSelectedNodeID = StockGV.GetRowCellValue(StockGV.FocusedRowHandle, "prod_category")
+            frmEditStock.ShowDialog()
+        End If
     End Sub
 
     Dim HideZeroStockState As Integer = 0 '0 = no click ; 1 = click
@@ -208,5 +233,15 @@ Public Class frmStockManagement
         e.Info.ToolTipType = ToolTipType.SuperTip
         e.Info.SuperTip = New SuperToolTip
         e.Info.SuperTip.Setup(toolTipArgs)
+    End Sub
+
+    Private Sub StockGV_RowStyle(ByVal sender As Object, ByVal e As DevExpress.XtraGrid.Views.Grid.RowStyleEventArgs) Handles StockGV.RowStyle
+        Dim view As GridView = sender
+        If (e.RowHandle >= 0) Then
+            Dim qty As Integer = view.GetRowCellDisplayText(e.RowHandle, view.Columns("prod_quantity"))
+            If qty <= 0 Then
+                e.Appearance.ForeColor = Color.DarkRed
+            End If
+        End If
     End Sub
 End Class
